+++ ewm-service/src/main/java/ru/practicum/ewm/adminApi/service/event/AdminEventsService.java
@@ -7,7 +7,7 @@ import ru.practicum.ewm.base.dto.UpdateEventAdminRequest;
 import java.util.List;
 
 public interface AdminEventsService {
-    EventFullDto update(Integer eventId, UpdateEventAdminRequest updateEvent);
+    EventFullDto update(Long eventId, UpdateEventAdminRequest updateEvent);
 
     List<EventFullDto> getAll(RequestParamForEvent param);
 }
