+++ ewm-service/src/main/java/ru/practicum/ewm/base/dto/UpdateEventUserRequest.java
@@ -14,14 +14,14 @@ import java.time.LocalDateTime;
 public class UpdateEventUserRequest {
     @Length(min = 20, max = 2000)
     private String annotation;
-    private Integer category;
+    private Long category;
     @Length(min = 20, max = 7000)
     private String description;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime eventDate;
     private LocationDto location;
     private boolean paid;
-    private int participantLimit;
+    private long participantLimit;
     private boolean requestModeration;
     private String stateAction;
     @Length(min = 3, max = 120)
