+++ ewm-service/src/main/java/ru/practicum/ewm/publicApi/controller/PublicEventsController.java
@@ -24,7 +24,7 @@ public class PublicEventsController {
 
     @GetMapping
     public ResponseEntity<List<EventShortDto>> getAll(@RequestParam(required = false) String text,
-                                                      @RequestParam(required = false) List<Integer> categories,
+                                                      @RequestParam(required = false) List<Long> categories,
                                                       @RequestParam(required = false) Boolean paid,
                                                       @RequestParam(required = false) String rangeStart,
                                                       @RequestParam(required = false) String rangeEnd,
@@ -52,7 +52,7 @@ public class PublicEventsController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<EventFullDto> get(@PathVariable Integer id) {
+    public ResponseEntity<EventFullDto> get(@PathVariable Long id) {
         log.info("Получен запрос GET /events/{}", id);
         return new ResponseEntity<>(eventsService.get(id), HttpStatus.OK);
     }
