+++ ewm-service/src/main/java/ru/practicum/ewm/publicApi/dto/RequestParamForEvent.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class RequestParamForEvent {
 
     private String text;
-    private List<Integer> categories;
+    private List<Long> categories;
     private Boolean paid;
     private String rangeStart;
     private String rangeEnd;
